requirements.txt
fastapi
uvicorn[standard]
chromadb
pydantic
sentence-transformers
ollama
ag2[openai]
python-dotenv
